+++ src/volatility/historical_vol.c
@@ -7,12 +7,12 @@
 double vol_close_to_close(const double *returns, int n)
 {
     double mean = 0.0;
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < (size_t)n; ++i)
         mean += returns[i];
     mean /= n;
 
     double var = 0.0;
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < (size_t)n; ++i)
         var += (returns[i] - mean) * (returns[i] - mean);
     var /= (n - 1);
 
@@ -23,7 +23,7 @@ double vol_close_to_close(const double *returns, int n)
 double vol_parkinson(const double *high, const double *low, int n)
 {
     double sum = 0.0;
-    for (int i = 0; i < n; ++i) 
+    for (size_t i = 0; i < (size_t)n; ++i) 
     {
         double hl = log(high[i] / low[i]);
         sum += hl * hl;
@@ -35,7 +35,7 @@ double vol_parkinson(const double *high, const double *low, int n)
 double vol_garman_klass(const double *high, const double *low, const double *open, const double *close, int n)
 {
     double sum = 0.0;
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < (size_t)n; ++i) {
         double hl = log(high[i] / low[i]);
         double co = log(close[i] / open[i]);
         sum += 0.5 * hl * hl - (2.0 * log(2.0) - 1.0) * co * co;
@@ -47,31 +47,31 @@ double vol_garman_klass(const double *high, const double *low, const double *ope
 double vol_yang_zhang(const double *high, const double *low, const double *open, const double *close, int n)
 {
     double mean_o = 0.0;
-    for (int i = 1; i < n; ++i)
+    for (size_t i = 1; i < (size_t)n; ++i)
         mean_o += log(open[i] / close[i - 1]);
     mean_o /= (n - 1);
 
     double var_o = 0.0;
-    for (int i = 1; i < n; ++i) {
+    for (size_t i = 1; i < (size_t)n; ++i) {
         double o = log(open[i] / close[i - 1]) - mean_o;
         var_o += o * o;
     }
     var_o /= (n - 2);
 
     double mean_c = 0.0;
-    for (int i = 0; i < n; ++i)
+    for (size_t i = 0; i < (size_t)n; ++i)
         mean_c += log(close[i] / open[i]);
     mean_c /= n;
 
     double var_c = 0.0;
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < (size_t)n; ++i) {
         double c = log(close[i] / open[i]) - mean_c;
         var_c += c * c;
     }
     var_c /= (n - 1);
 
     double var_rs = 0.0;
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < (size_t)n; ++i) {
         double ho = log(high[i] / open[i]);
         double hc = log(high[i] / close[i]);
         double lo = log(low[i] / open[i]);
